+++ src/App.css
@@ -7,7 +7,7 @@
   margin: auto;
 }
 
-@media(min-width: 1280px){
+/*@media(min-width: 1280px){
   .App{
     background-image: url(koro.jpg);
     background-repeat: no-repeat;
@@ -25,7 +25,7 @@
     background-position: center;
     width: 80%;
   }
-}
+}*/
 
 .App-logo {
   height: 40vmin;
@@ -42,11 +42,9 @@
   color: rgb(0, 0, 0);
 }
 
-.App-chatroom {
+.App-content {
   flex: 12;
-
   background-color: rgba(201, 137, 197, 0.418);
-
 }
 
 .App-chatroom-text {
