+++ src/App.js
@@ -1,27 +1,8 @@
 import './App.css';
-import { useState } from 'react';
 import { Link } from 'react-router-dom';
 
-function App() {
 
-    const [text,setText] = useState("");
-
-    const [lines,setLines] = useState([]);
-
-    const onTextChange = (event) => {
-      setText(event.target.value);
-    };
-
-    const onSend =() => {
-      setLines(lines => [...lines,text]);
-      setText("");
-    };
-
-    const keyPress = (event) => {
-      if (event.which === 13){
-        onSend();
-      }
-    };
+function App({children}) {
 
   return (
     <div className="App">
@@ -32,21 +13,9 @@ function App() {
                   </button>
               </Link>
         </div>
-      <div className="App-chatroom">
-        {
-          lines.map(x =>{
-           return <div className="App-chatroom-text">
-                      {x}
-                  </div>
-          })
-        } 
-      </div>
-      <div className="App-textbox">
-      <input type="text" className="App-textbox-input" 
-        value={text} onChange={onTextChange} onKeyPress={keyPress}/>
-
-        <div className="App-textbox-send" onClick={onSend}>Send!</div>
-      </div>
+        <div className="App-content">
+          {children}
+        </div>
     </div>
   );
 }
