+++ src/Chatbox.css
@@ -1,0 +1,103 @@
+.App {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    text-align: center;
+    height: 100%;
+    margin: auto;
+  }
+  
+  /*@media(min-width: 1280px){
+    .App{
+      background-image: url(koro.jpg);
+      background-repeat: no-repeat;
+      background-position: center;
+      width:95%; 
+      height:95%;
+    }
+  }
+  
+  @media (max-width: 768px) {
+    .App {
+      background-color: rgb(120, 193, 221);
+      background-image: url(Hol.png);
+      background-repeat: no-repeat;
+      background-position: center;
+      width: 80%;
+    }
+  }*/
+  
+  .App-logo {
+    height: 40vmin;
+    pointer-events: none;
+  }
+  
+  .App-header {
+    flex: 1;
+    background-color: #121079;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    color: rgb(0, 0, 0);
+  }
+  
+  .App-chatroomlab {
+    flex: 12;
+    flex-direction: column;
+    background-color: rgba(201, 137, 197, 0.418);
+    background-image: url(koro.jpg);
+    background-repeat: no-repeat;
+    background-position: center;
+  }
+  
+  .App-chatroom-text {
+    margin: 8px;
+  
+    background-color: rgb(109, 196, 120);
+  
+    color : rgb(48, 100, 100);
+    font-size: 20px;
+  }
+  
+  .App-textbox {
+    flex: 1;
+    background-color: rgb(70, 62, 94);
+    justify-content: center;
+    align-items: stretch;
+    display: flex;
+  }
+  
+  .App-textbox-input {
+    flex: 5;
+    height: 100%;
+    box-sizing: border-box;
+  }
+  
+  .App-textbox-send {
+    flex: 1;
+    height: 100%;
+    vertical-align: middle;
+    justify-content: center;
+    align-items: center;
+    display: flex;
+  }
+  
+  .col-1 {width:  8.33%;}
+  .col-2 {width:  16.66%;}
+  .col-3 {width:  25%;}
+  .col-4 {width:  33.33%;}
+  .col-5 {width:  41.66%;}
+  .col-6 {width:  50%;}
+  .col-7 {width:  58.33%;}
+  .col-8 {width:  66.66%;}
+  .col-9 {width:  75%;}
+  .col-10 {width:  83.33%;}
+  .col-11{width:  91.66%;}
+  .col-12 {width: 100%;}
+  
+  @media (max-width:786px){
+    [class*="col-"]{
+      width: 100;
+    }
+  }
