+++ src/Router.js
@@ -1,7 +1,8 @@
 import {BrowserRouter, Switch , Route} from 'react-router-dom';
-/*import App from './App';*/
+import Chatbox from './Chatbox';
 import ChatRoom from './ChatRoom';
 import ProFriend from './project/ProFriend';
+import FriendList from './FriendList';
 
 
 function Router() {
@@ -12,8 +13,14 @@ function Router() {
                 <Route path="/chatroom">
                     <ChatRoom />
                 </Route>
+                <Route path="/fri">
+                    <FriendList />
+                </Route>
                 <Route path="/ProFriend">
                     <ProFriend />
+                </Route>
+                <Route path="/">
+                    <Chatbox />
                 </Route>
             </Switch>
         </BrowserRouter>
